Outputs 1s and 0s with incorrect timing/frequency
Troubleshooting timing issues.  Today I got 1s and 0s to display on screen after eight taps of 1 or 0 but only one comes up after 8 taps, not 8 numbers, but one.  Sometimes the first time around it’s 4 digits that come up at once but thereafter only 1 at a time.  Timing is off, may have to reinstate state machine:

In simulation I can never get the right timing for the bit reversal tx_data signal (this is the output to the screen). Its one of two options depending on if I include an else statement in the alwyas @* block in message_ram or not.